pyproject.toml
[project]
name = "test1"
version = "0.0.1"
description = "Description for test1 project"
authors = [{ name = "John Doe", email = "john.doe@myemail.com" }]
dependencies = [
    "uipath-mcp>=0.0.106",
]
requires-python = ">=3.11"
